refactor(types): move observable matcher types back to grafana-test-utils
The declare global jest augmentation for toEmitValues/toEmitValuesWith belongs
in grafana-test-utils alongside its implementation — types flow to consumers
via the import chain rather than needing a separate globals declaration.

Removes observable-matchers.d.ts from @types/grafana-globals and restores
the declare global block in grafana-test-utils/src/matchers/types.ts.
