+++ packages/grafana-global-types/index.d.ts
@@ -3,4 +3,3 @@ export * from './images';
 export * from './intl';
 export * from './jquery';
 export * from './jest-dom';
-export * from './observable-matchers';
